+++ packages/compose/src/hooks/use-focus-outside/test/index.js
@@ -9,8 +9,6 @@ import userEvent from '@testing-library/user-event';
  */
 import useFocusOutside from '../';
 
-jest.useFakeTimers();
-
 const FocusOutsideComponent = ( { onFocusOutside: callback } ) => (
 	<div>
 		{ /* Wrapper */ }
@@ -29,9 +27,7 @@ const FocusOutsideComponent = ( { onFocusOutside: callback } ) => (
 describe( 'useFocusOutside', () => {
 	it( 'should not call handler if focus shifts to element within component', async () => {
 		const mockOnFocusOutside = jest.fn();
-		const user = userEvent.setup( {
-			advanceTimers: jest.advanceTimersByTime,
-		} );
+		const user = userEvent.setup();
 
 		render(
 			<FocusOutsideComponent onFocusOutside={ mockOnFocusOutside } />
@@ -52,9 +48,7 @@ describe( 'useFocusOutside', () => {
 
 	it( 'should not call handler if focus transitions via click to button', async () => {
 		const mockOnFocusOutside = jest.fn();
-		const user = userEvent.setup( {
-			advanceTimers: jest.advanceTimersByTime,
-		} );
+		const user = userEvent.setup();
 
 		render(
 			<FocusOutsideComponent onFocusOutside={ mockOnFocusOutside } />
@@ -71,9 +65,7 @@ describe( 'useFocusOutside', () => {
 
 	it( 'should call handler if focus shifts to element outside component', async () => {
 		const mockOnFocusOutside = jest.fn();
-		const user = userEvent.setup( {
-			advanceTimers: jest.advanceTimersByTime,
-		} );
+		const user = userEvent.setup();
 
 		render(
 			<FocusOutsideComponent onFocusOutside={ mockOnFocusOutside } />
@@ -101,9 +93,7 @@ describe( 'useFocusOutside', () => {
 			.spyOn( document, 'hasFocus' )
 			.mockImplementation( () => false );
 		const mockOnFocusOutside = jest.fn();
-		const user = userEvent.setup( {
-			advanceTimers: jest.advanceTimersByTime,
-		} );
+		const user = userEvent.setup();
 
 		render(
 			<FocusOutsideComponent onFocusOutside={ mockOnFocusOutside } />
@@ -128,9 +118,7 @@ describe( 'useFocusOutside', () => {
 
 	it( 'should cancel check when unmounting while queued', async () => {
 		const mockOnFocusOutside = jest.fn();
-		const user = userEvent.setup( {
-			advanceTimers: jest.advanceTimersByTime,
-		} );
+		const user = userEvent.setup();
 
 		const { unmount } = render(
 			<FocusOutsideComponent onFocusOutside={ mockOnFocusOutside } />
@@ -146,8 +134,6 @@ describe( 'useFocusOutside', () => {
 		// handler is queued
 		button.blur();
 		unmount();
-
-		jest.runAllTimers();
 
 		expect( mockOnFocusOutside ).not.toHaveBeenCalled();
 	} );