+++ packages/edit-site/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Bug Fix
+
+-   Force visual editor in browse mode ([#47329](https://github.com/WordPress/gutenberg/pull/47329)).
+
+## 5.2.0 (2023-01-11)
+
 ## 5.1.0 (2023-01-02)
 
 ## 5.0.0 (2022-12-14)