+++ packages/react-native-editor/CHANGELOG.md
@@ -11,6 +11,17 @@ For each user feature we should also add a importance categorization label  to i
 
 ## Unreleased
 
+## 1.87.3
+-   [*] Fix insert blocks not handling raw string properly in unsupported block editor [#47472]
+
+## 1.87.2
+-   [*] Add boolean contentStyle and clientId check to Column Edit InnerBlocks [#47234]
+-   [*] Line-height and font-size regression fixes [#47284]
+
+## 1.87.1
+-   [**] Gallery block: Address styling regression, in which negative margin was added [#47086]
+-   [*] RichText - Parse CSS values and avoid setting undefined ones [#47080]
+
 ## 1.87.0
 -   [*] Add capabilities to force only Core blocks and control Support section [#46215]